fix: resolve compile issues and harden auction domain logic
Agent-Logs-Url: https://github.com/v1-log/end_year/sessions/b2274651-7adc-445c-8b4a-309bde636190

+++ Project_36/src/main/java/com/auction/model/ItemType/Art.java
@@ -1,7 +1,9 @@
 package com.auction.model.ItemType;
 
+import java.math.BigDecimal;
+
 public class Art extends Item {
-    public Art(String name, double price) {
+    public Art(String name, BigDecimal price) {
         super(name, price);
     }
 }
