+++ Project_36/src/main/java/com/auction/model/ItemType/Electronics.java
@@ -1,7 +1,9 @@
 package com.auction.model.ItemType;
 
+import java.math.BigDecimal;
+
 public class Electronics extends Item {
-    public Electronics(String name, double price) {
+    public Electronics(String name, BigDecimal price) {
         super(name, price);
     }
 }
