+++ Project_36/src/main/java/com/auction/model/ItemType/Vehicle.java
@@ -1,7 +1,9 @@
 package com.auction.model.ItemType;
 
+import java.math.BigDecimal;
+
 public class Vehicle extends Item {
-    public Vehicle(String name, double price) {
+    public Vehicle(String name, BigDecimal price) {
         super(name, price);
     }
 }
